update github logiin
